+++ ckanext/opendata_theme/base/fanstatic/js/color_picker.js
