+++ ckanext/opendata_theme/base/fanstatic/resource.config
