+++ ckanext/opendata_theme/opengov_custom_css/templates/base.html
@@ -3,7 +3,7 @@
 {% block styles %}
   {{ super() }}
     <style>
-        {% resource 'opendata_custom_css_resource/css/theme.css' %}
+        {% resource 'opengov_custom_css_resource/css/theme.css' %}
         {{ g['ckanext.opendata_theme.custom_raw_css'] | safe }}
     </style>
 {% endblock %}
