+++ ckanext/opendata_theme/opengov_custom_homepage/templates/home/snippets/search.html
