--- ckanext/opendata_theme/tests/test_plugin.py
@@ -1,5 +1,0 @@
-"""Tests for plugin.py."""
-import ckanext.opendata_theme.plugin as plugin
-
-def test_plugin():
-    pass
